PR go/69511     runtime: change G gcstack_size field to size_t
    Because its address is passed to __splitstack_find, which expects size_t*.
    
    From Dominik Vogt in GCC PR 69511.
    
    Reviewed-on: https://go-review.googlesource.com/19429


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@233260 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/runtime/proc.c
@@ -2267,7 +2267,7 @@ runtime_malg(int32 stacksize, byte** ret_stack, size_t* ret_stacksize)
 		}
 		*ret_stacksize = stacksize;
 		newg->gcinitial_sp = *ret_stack;
-		newg->gcstack_size = stacksize;
+		newg->gcstack_size = (size_t)stacksize;
 #endif
 	}
 	return newg;
